==> Final Code/2. Pull Combine Data.ipynb <==
# Setup 

import os, time 
import pandas as pd 
import numpy as np 

    # get the file paths from environment variables 
folder_path = os.environ.get("NFL_DATA_PATH") 

# turn off the pandas warning 
pd.options.mode.chained_assignment = None  # default='warn' 

# Get Player List 

# loop through each week and process the data 
df_players = pd.DataFrame() 
for weeknum in range(1, 18): 
    print(f"Processing week {weeknum}") 
    
    # read in the data 
    df = pd.read_csv(f"{folder_path}//train//input_2023_w{weeknum:02}.csv") 

    # subset to the player metadata 
    df = df[["nfl_id", "player_name", "player_position"]].drop_duplicates().reset_index(drop=True) 

    # append to the overall dataframe 
    df_players = (
        pd.concat([df_players, df], axis=0, ignore_index=True) 
        .drop_duplicates().reset_index(drop=True) 
    ) 

# clean up a few of the columns 
df_players["player_name_upper"] = df_players["player_name"].str.upper() 

# save to a csv 
df_players.to_csv(f"{folder_path}//player_metadata.csv", index=False) 

# showcase the data 
df_players.head() 

# Pull Combine Data 

# loop through each year and add it 
df_all = pd.DataFrame() 
for year in range(2010, 2023): 
    print(f"Processing year {year}") 

    # define the table link
    table_link = f"https://www.pro-football-reference.com/draft/{year}-combine.htm#combine" 

    # download the table 
    df_combine = pd.read_html(table_link)[0] 

    # add the year column 
    df_combine["draft_year"] = year 

    # filter the data 
    df_combine = (
        df_combine.loc[df_combine["40yd"].notna()] 
        [["Player", "40yd", "3Cone", "Shuttle", "draft_year"]]
    ) 

    # append to the overall dataframe 
    df_all = pd.concat([df_all, df_combine], axis = 0, ignore_index = True) 

    # add a short delay to be polite to the server 
    time.sleep(1) 

# correct the column types 
df_all["40yd"] = pd.to_numeric(df_all["40yd"], errors = "coerce") 
df_all["3Cone"] = pd.to_numeric(df_all["3Cone"], errors = "coerce")
df_all["Shuttle"] = pd.to_numeric(df_all["Shuttle"], errors = "coerce") 

# showcase the data 
df_all.head() 

# Join and Save 

# join in the NFL ids 
df_all["player_name_upper"] = df_all["Player"].str.upper() 
df_players["player_name_upper"] = df_players["player_name_upper"].apply(lambda x: x.replace(" II", ""))
df_full = df_all.merge(df_players, on = "player_name_upper", how = "inner") 

# get the latest stat for each player 
df_full = df_full.sort_values(["player_name_upper", "draft_year"]).drop_duplicates("player_name_upper", keep = "last") 

# save to csv 
df_full.to_csv(f"{folder_path}//combine_data.csv", index = False) 

# showcase the data 
print(f"Total Players with 40 times: {len(df_all.index):,}")
df_all.head()
